requirements.txt
# minimal requirements (tested with Python 3.10+ but works with 3.11)
libclang
neo4j==5.28.0
ollama>=0.4.0
streamlit>=1.18.0
python-dotenv>=1.0.0
ujson>=5.8.0
